backend/requirements.txt
fastapi
uvicorn
requests
pydantic
PyYAML
langchain
langchain-ollama
